refactor files and packages
