+++ app/src/main/java/edu/up/cs/androidcatan/catan/MainActivity.java
@@ -1,4 +1,4 @@
-package edu.up.cs.androidcatan;
+package edu.up.cs.androidcatan.catan;
 
 import android.view.View;
 
@@ -9,7 +9,7 @@ import edu.up.cs.androidcatan.game.GamePlayer;
 import edu.up.cs.androidcatan.game.LocalGame;
 import edu.up.cs.androidcatan.game.config.GameConfig;
 import edu.up.cs.androidcatan.game.config.GamePlayerType;
-import edu.up.cs.androidcatan.players.Player;
+import edu.up.cs.androidcatan.catan.players.*;
 
 /**
  * @author Alex Weininger
