+++ app/src/main/java/edu/up/cs/androidcatan/catan/graphics/HexagonDrawable.java
@@ -1,4 +1,4 @@
-package edu.up.cs.androidcatan.catan.hexagon;
+package edu.up.cs.androidcatan.catan.graphics;
 
 import android.content.Context;
 import android.graphics.Canvas;
@@ -8,9 +8,6 @@ import android.graphics.Path;
 import android.util.Log;
 
 import java.util.Random;
-
-import edu.up.cs.androidcatan.catan.graphics.RoadDrawable;
-import edu.up.cs.androidcatan.catan.graphics.boardSurfaceView;
 
 public class HexagonDrawable extends boardSurfaceView {
